feat: Implement adding monthly repetition

+++ src/main/java/ru/steqa/repetitionservice/scheme/rabbit/repetition/FixedMonthRepetition.java
@@ -7,10 +7,10 @@ import ru.steqa.repetitionservice.scheme.rabbit.TransactionType;
 import java.time.LocalDateTime;
 
 public class FixedMonthRepetition extends BaseRepetitionRule {
-    public Byte day;
+    public Integer day;
 
     public FixedMonthRepetition(Long userId, Long transactionId, TransactionType transactionType,
-                                LocalDateTime nextExecution, Byte day) {
+                                LocalDateTime nextExecution, Integer day) {
         super(userId, transactionId, transactionType, nextExecution, RepetitionMode.FIXED_MONTH);
         this.day = day;
     }
